[fix] MSDA 연산 시 fp32로 강제 캐스팅

+++ models/dn_dab_dino_deformable_detr/ops/functions/ms_deform_attn_func.py
@@ -17,13 +17,12 @@ from torch.autograd.function import once_differentiable
 
 import MultiScaleDeformableAttention as MSDA
 
-
 class MSDeformAttnFunction(Function):
     @staticmethod
     def forward(ctx, value, value_spatial_shapes, value_level_start_index, sampling_locations, attention_weights, im2col_step):
         ctx.im2col_step = im2col_step
         output = MSDA.ms_deform_attn_forward(
-            value, value_spatial_shapes, value_level_start_index, sampling_locations, attention_weights, ctx.im2col_step)
+            value.type(torch.float32), value_spatial_shapes, value_level_start_index, sampling_locations, attention_weights, ctx.im2col_step)
         ctx.save_for_backward(value, value_spatial_shapes, value_level_start_index, sampling_locations, attention_weights)
         return output
